use f64 for projection matrices and reduce fNear parameter

+++ cubetracer/shaders/global_uniforms.h
@@ -8,9 +8,9 @@ layout(set=G_UNIFORM_SET, binding = 0) uniform accelerationStructureNV UNI_TLAS;
 #endif
 
 layout(set=G_UNIFORM_SET, binding = 1) uniform SCamera{
-    mat4 screen_to_world;
-    mat4 prev_world_to_screen;
-    vec3 origin;
+    dmat4 screen_to_world;
+    dmat4 prev_world_to_screen;
+    dvec3 origin;
 } UNI_CAMERA;
 
 layout(set=G_UNIFORM_SET, binding = 2) uniform SScene {
